chore(webapp): rename /service to /api
