+++ webapp/javascript/components/exportToFlamegraphDotCom.hook.ts
@@ -1,5 +1,5 @@
 import { Profile } from '@pyroscope/models/src';
-import { shareWithFlamegraphDotcom } from '@webapp/services/share';
+import { shareWithFlamegraphDotcom } from '@webapp/api/share';
 import { useAppDispatch } from '@webapp/redux/hooks';
 import handleError from '@webapp/util/handleError';
